% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_configs.R
\docType{data}
\name{DOMAIN_CONFIGS}
\alias{DOMAIN_CONFIGS}
\title{Domain Configurations Based on OMOP CDM Tables}
\format{
A list of lists where each sublist contains:
\describe{
\item{table}{Table name in the OMOP CDM}
\item{date_field}{Single date field name, if applicable}
\item{start_field}{Start date field name, if applicable}
\item{end_field}{End date field name, if applicable}
}
}
\usage{
DOMAIN_CONFIGS
}
\description{
A list containing configurations for each domain in the OMOP CDM, including
table names and relevant date fields.
}
\keyword{datasets}
