% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescription_rates.R
\name{get_atc_patient_count}
\alias{get_atc_patient_count}
\title{Get Patient Count for ATC Category}
\usage{
get_atc_patient_count(conn, cdm_schema, atc_code)
}
\arguments{
\item{conn}{A DBI connection object to the database}

\item{cdm_schema}{Character string specifying the CDM schema name}

\item{atc_code}{Character string specifying the ATC category (first letter)}
}
\value{
Integer representing the count of unique patients
}
\description{
Calculates the number of unique patients who received prescriptions in a specific ATC category
}
\examples{
\dontrun{
conn <- DBI::dbConnect(...)
count <- get_atc_patient_count(conn, "cdm_schema", "C")
}
}
