% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescription_rates.R
\name{get_total_prescriptions}
\alias{get_total_prescriptions}
\title{Get Total Prescription Count}
\usage{
get_total_prescriptions(conn, cdm_schema)
}
\arguments{
\item{conn}{A DBI connection object to the database}

\item{cdm_schema}{Character string specifying the CDM schema name}
}
\value{
Integer representing the total count of prescriptions
}
\description{
Calculates the total number of prescriptions in the database
}
\examples{
\dontrun{
conn <- DBI::dbConnect(...)
total <- get_total_prescriptions(conn, "cdm_schema")
}
}
