% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_domain}
\alias{validate_domain}
\title{Validate Domain Name}
\usage{
validate_domain(domain)
}
\arguments{
\item{domain}{Character string of domain name}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate domain name
}
\keyword{internal}
